+++ src/client/java/dinner/dev/endless_gravity/EndlessGravityConfigScreen.java
@@ -124,6 +124,8 @@ public class EndlessGravityConfigScreen {
             addToggle(atmosphere, entry, "endless_gravity.config.overworldEntityGravity", Config.COMMON.overworldEntityGravity, true);
             addDouble(atmosphere, entry, "endless_gravity.config.atmosphereGravityMax", Config.COMMON.atmosphereGravityMax,
                     0.0, 0.1, 0.075, TEXT_WHITE);
+            addDouble(atmosphere, entry, "endless_gravity.config.noFallDamageAltitude", Config.COMMON.noFallDamageAltitude,
+                    64.0, 3500.0, 400.0, TEXT_WHITE);
             addDouble(atmosphere, entry, "endless_gravity.config.atmosphereMuffleGain", Config.COMMON.atmosphereMuffleGain,
                     0.0, 1.0, 0.01, TEXT_WHITE);
             addDouble(atmosphere, entry, "endless_gravity.config.atmosphereMuffleGainHF", Config.COMMON.atmosphereMuffleGainHF,